Now records gyro and accel data.
Does __NOT__ record touch events. That's a little trickier to get
working, because it's not something that can be done in
the background.

Also, JSONKit was added as a dependency but is not used. Instead
we're converting the NSMutableArray objects for the gyro and accel
data to JSON using Apple's libraries.

Not tested because I don't have a certificate.

	new file:   .gitmodules
	modified:   SampleApp/SAAppDelegate.h
	modified:   SampleApp/SAAppDelegate.m
	modified:   SampleApp/SAViewController.m
	new file:   SampleApp/libs/JSONKit